Refactor: Extract code into separate DocumentManipulationService
