+++ XamlStyler.Service/DocumentManipulation/IProcessElementService.cs
@@ -1,6 +1,6 @@
 using System.Xml.Linq;
 
-namespace XamlStyler.Core.Reorder
+namespace XamlStyler.Core.DocumentManipulation
 {
     public interface IProcessElementService
     {
