+++ XamlStyler.Service/DocumentManipulation/ThicknessFormatter.cs
@@ -1,7 +1,7 @@
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace XamlStyler.Core.Reorder
+namespace XamlStyler.Core.DocumentManipulation
 {
     public static class ThicknessFormatter
     {
