+++ XamlStyler.Service/Extensions/StringExtension.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Linq;
 using System.Text;
-using XamlStyler.Core.Reorder;
+using XamlStyler.Core.DocumentManipulation;
 
 namespace XamlStyler.Core.Extensions
 {
