+++ XamlStyler.Service/Options/IStylerOptions.cs
@@ -1,4 +1,4 @@
-using XamlStyler.Core.Reorder;
+using XamlStyler.Core.DocumentManipulation;
 
 namespace XamlStyler.Core.Options
 {
