+++ XamlStyler.Service/XamlStyler.Service.csproj
@@ -47,6 +47,7 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DocumentManipulation\DocumentManipulationService.cs" />
     <Compile Include="MarkupExtensions\Formatter\AttributeInfoFactory.cs" />
     <Compile Include="MarkupExtensions\Formatter\MarkupExtensionFormatterBase.cs" />
     <Compile Include="Extensions\XmlReaderExtensions.cs" />
@@ -70,18 +71,18 @@
     <Compile Include="MarkupExtensions\Parser\MemberNameOrStringTerminal.cs" />
     <Compile Include="MarkupExtensions\Parser\TypeNameTerminal.cs" />
     <Compile Include="MarkupExtensions\Parser\XamlMarkupExtensionGrammar.cs" />
-    <Compile Include="Reorder\IProcessElementService.cs" />
-    <Compile Include="Reorder\ISortableAttribute.cs" />
-    <Compile Include="Reorder\NameSelector.cs" />
-    <Compile Include="Reorder\NodeCollection.cs" />
-    <Compile Include="Reorder\NodeReorderService.cs" />
-    <Compile Include="Reorder\FormatThicknessService.cs" />
-    <Compile Include="Reorder\SortableNumericAttribute.cs" />
-    <Compile Include="Reorder\SortableStringAttribute.cs" />
-    <Compile Include="Reorder\SortBy.cs" />
-    <Compile Include="Reorder\ThicknessFormatter.cs" />
-    <Compile Include="Reorder\ThicknessStyle.cs" />
-    <Compile Include="Reorder\Wildcard.cs" />
+    <Compile Include="DocumentManipulation\IProcessElementService.cs" />
+    <Compile Include="DocumentManipulation\ISortableAttribute.cs" />
+    <Compile Include="DocumentManipulation\NameSelector.cs" />
+    <Compile Include="DocumentManipulation\NodeCollection.cs" />
+    <Compile Include="DocumentManipulation\NodeReorderService.cs" />
+    <Compile Include="DocumentManipulation\FormatThicknessService.cs" />
+    <Compile Include="DocumentManipulation\SortableNumericAttribute.cs" />
+    <Compile Include="DocumentManipulation\SortableStringAttribute.cs" />
+    <Compile Include="DocumentManipulation\SortBy.cs" />
+    <Compile Include="DocumentManipulation\ThicknessFormatter.cs" />
+    <Compile Include="DocumentManipulation\ThicknessStyle.cs" />
+    <Compile Include="DocumentManipulation\Wildcard.cs" />
     <Compile Include="Options\IStylerOptions.cs" />
     <Compile Include="Options\LineBreakRule.cs" />
     <Compile Include="Options\ReorderSettersBy.cs" />
